+++ index.php
@@ -13,23 +13,23 @@ require_once 'src/Autoloader.php';
 Autoloader::register();
 
 // Check IP, deny access if not allowed
-if(!(empty(Config::ACCESS_IP) OR $_SERVER['REMOTE_ADDR'] == "127.0.0.1" OR $_SERVER['REMOTE_ADDR'] == "::1" OR $_SERVER['REMOTE_ADDR'] == Config::ACCESS_IP OR $_SERVER['REMOTE_ADDR'] == Config::ACCESS_IP2 OR $_SERVER['REMOTE_ADDR'] == Config::ACCESS_IP3)){
+if(!(empty(Config::ACCESS_IP) OR $_SERVER['REMOTE_ADDR'] == '127.0.0.1' OR $_SERVER['REMOTE_ADDR'] == '::1' OR $_SERVER['REMOTE_ADDR'] == Config::ACCESS_IP OR $_SERVER['REMOTE_ADDR'] == Config::ACCESS_IP2 OR $_SERVER['REMOTE_ADDR'] == Config::ACCESS_IP3)){
 	header('Location: login.html');
 	exit; 
 }
 
 // Start check user session
 session_start();
-$passToken = hash('sha256', Config::PASSWORD."ibe81rn6");
+$passToken = hash('sha256', Config::PASSWORD.'ibe81rn6');
 
 // Active Session
 if(isset($_SESSION['login']) AND $_SESSION['login'] === TRUE){
 	// Nothing needs to be done
 	
 // Login Cookie available	
-}elseif(isset($_COOKIE["Login"]) AND $_COOKIE["Login"] == $passToken){
+}elseif(isset($_COOKIE['Login']) AND $_COOKIE['Login'] == $passToken){
 		$_SESSION['login'] = TRUE;
-		$_SESSION["csfrToken"] = hash('sha256', random_bytes(20));
+		$_SESSION['csfrToken'] = hash('sha256', random_bytes(20));
 
 // Login		
 }elseif(!isset($_SESSION['login']) AND isset($_POST['password']) AND $_POST['password'] == Config::PASSWORD){
@@ -37,9 +37,9 @@ if(isset($_SESSION['login']) AND $_SESSION['login'] === TRUE){
 	$passHashed = hash('sha256', Config::PASSWORD);
 	
 		$_SESSION['login'] = TRUE;
-		$_SESSION["csfrToken"] = hash('sha256', random_bytes(20));
+		$_SESSION['csfrToken'] = hash('sha256', random_bytes(20));
 		if(isset($_POST['stayloggedin'])){		
-			setcookie("Login", $passToken, time()+2592000, "","",FALSE, TRUE);
+			setcookie('Login', $passToken, time()+2592000, '','',FALSE, TRUE);
 		}
 
 // Not logged in or invalid data
@@ -48,42 +48,87 @@ if(isset($_SESSION['login']) AND $_SESSION['login'] === TRUE){
 //	exit; 	
 }
 
-// Load utiliy and content creator functions
+// Load utility and content creator functions
 require_once 'src/Utility.php';
 require_once 'src/Content.php';
 require_once __DIR__ . '/vendor/autoload.php';
 
+use SQLite3;
+
 // Globals
-$error = "";
-$message = "";
+$error = '';
+$message = '';
 $trafficC = 0;
 $trafficCIn = 0;
 $trafficCOut = 0;
 $newPeersCount = 0;
 $coind = new jsonRPCClient('http://'.Config::RPC_USER.':'.Config::RPC_PASSWORD.'@'.Config::RPC_IP.'/', Config::DEBUG);
 $coinApi = new \Coinpaprika\Client();
+$db = new SQLite3('data/scalaris.db');
+
+// Do some database setup
+$db->enableExceptions(true);
+
+$db->exec('CREATE TABLE IF NOT EXISTS "pastorders"(
+	"id" VARCHAR PRIMARY KEY NOT NULL,
+	"timestamp" INTEGER NOT NULL,
+	"fee_txid" VARCHAR NOT NULL,
+	"nodepubkey" VARCHAR NOT NULL,
+	"taker" VARCHAR NOT NULL,
+	"taker_size" INTEGER NOT NULL,
+	"maker" VARCHAR NOT NULL,
+	"maker_size" INTEGER NOT NULL
+)');
+
+$db->exec('CREATE TABLE IF NOT EXISTS "events"(
+	"lastorderheight" INTEGER,
+	"lastproposal" INTEGER,
+	"timestamp" INTEGER
+)');
+
+$db->exec('CREATE TABLE IF NOT EXISTS "pastproposals"(
+	"hash" VARCHAR PRIMARY KEY NOT NULL,
+	"name" VARCHAR NOT NULL,
+	"superblock" INTEGER NOT NULL,
+	"amount" INTEGER NOT NULL,
+	"address" VARCHAR NOT NULL,
+	"url" VARCHAR NOT NULL,
+	"description" VARCHAR,
+	"yeas" INTEGER NOT NULL,
+	"nays" INTEGER NOT NULL,
+	"abstains" INTEGER NOT NULL,
+	"status" VARCHAR NOT NULL
+)');
+
+$db->exec('CREATE TABLE IF NOT EXISTS "scratch"(
+   "superblock" VARCHAR NOT NULL,
+   "txid" VARCHAR NOT NULL,
+   "amount" INTEGER NOT NULL,
+   "address" VARCHAR NOT NULL
+)');
+
 
 // Content
 // Main Page
-if(empty($_GET) OR $_GET['p'] == "main") {   
+if(empty($_GET) OR $_GET['p'] == 'main') {   
 	try{
 	$content = createMainContent();
 	}catch(\Exception $e) {
-	   $error = "Node offline or incorrect RPC data";
+	   $error = 'Node offline or incorrect RPC data';
 	}
 	$data = array('section' => 'main', 'title' => 'Home', 'content' => $content);   
    
 // New Main Page
-}elseif($_GET['p'] == "newmain") {   
+}elseif($_GET['p'] == 'newmain') {   
 	try{
 	$content = createMainContent();
 	}catch(\Exception $e) {
-	   $error = "Node offline or incorrect RPC data";
+	   $error = 'Node offline or incorrect RPC data';
 	}
 	$data = array('section' => 'newmain', 'title' => 'Home', 'content' => $content);   
    
 // Peers Page   
-}elseif($_GET['p'] == "peers") {
+}elseif($_GET['p'] == 'peers') {
 	
 	// Information for header
 	$content = createPeerContent();
@@ -92,7 +137,7 @@ if(empty($_GET) OR $_GET['p'] == "main") {
 	$data = array('section' => 'peers', 'title' => 'Peers', 'content' => $content);
 
 // Memory Pool Page	
-}elseif($_GET['p'] == "mempool") {
+}elseif($_GET['p'] == 'mempool') {
 	
 	if(isset($_GET['e']) AND ctype_digit($_GET['id'])){
 		$end = $_GET['e'];
@@ -105,83 +150,109 @@ if(empty($_GET) OR $_GET['p'] == "main") {
  
  
 // Servicenodes Page
-}elseif($_GET['p'] == "servicenodes") {
-	$content = createNodesContent();
+}elseif($_GET['p'] == 'servicenodes') {
+	$content = createSNodesContent();
 	$data = array('section' => 'servicenodes', 'title' => 'Servicenodes', 'content' => $content);  
  
 // Proposals Page
-}elseif($_GET['p'] == "proposals") {
+}elseif($_GET['p'] == 'proposals') {
 	$content = createGovernanceContent();
 	$data = array('section' => 'proposals', 'title' => 'Proposals', 'content' => $content);
 
 // Past proposals Page
-}elseif($_GET['p'] == "pastproposals") {
-	$content = createOldGovernanceContent();
+}elseif($_GET['p'] == 'pastproposals') {
+	$content = createPastProposalsContent();
 	$data = array('section' => 'pastproposals', 'title' => 'Past Proposals', 'content' => $content);
 
 // Blocks Page 
-}elseif($_GET['p'] == "blocks") {
+}elseif($_GET['p'] == 'blocks') {
 	$content= createBlocksContent();
 	$data = array('section' => 'blocks', 'title' => 'Blocks', 'content' => $content);
   
 // Forks Page 
-}elseif($_GET['p'] == "forks") {
+}elseif($_GET['p'] == 'forks') {
 	$content= createForksContent();
 	$data = array('section' => 'forks', 'title' => 'Forks', 'content' => $content);
   
 // Open Orders Page 
-}elseif($_GET['p'] == "openorders") {
+}elseif($_GET['p'] == 'openorders') {
 	$content= createOpenOrdersContent();
 	$data = array('section' => 'openorders', 'title' => 'Open Orders', 'content' => $content);
   
 // Past Orders Page 
-}elseif($_GET['p'] == "past1") {
-	$content= createPastOrdersContent(1);
-	$data = array('section' => 'pastorders', 'title' => 'Past Orders', 'content' => $content);
-  
-// Past Orders Page 
-}elseif($_GET['p'] == "past7") {
-	$content= createPastOrdersContent(7);
-	$data = array('section' => 'pastorders', 'title' => 'Past Orders', 'content' => $content);
-  
-// Past Orders Page 
-}elseif($_GET['p'] == "past14") {
-	$content= createPastOrdersContent(14);
-	$data = array('section' => 'pastorders', 'title' => 'Past Orders', 'content' => $content);
-  
-// Past Orders Page 
-}elseif($_GET['p'] == "past30") {
-	$content= createPastOrdersContent(30);
-	$data = array('section' => 'pastorders', 'title' => 'Past Orders', 'content' => $content);
-  
-// Settings Page	
-}elseif($_GET['p'] == "settings") {
-	if(isset($_GET['c'])  AND $_GET['t'] == $_SESSION["csfrToken"]){
-		if(isset($_GET['c']) AND $_GET['c'] == "geosave"){
-			// Check if Geo Peer Tracing was changed
-			if(isset($_POST['geopeers']) AND $_POST['geopeers'] == "on"){
-				 $geoPeers = "true";
-			}else{
-				$geoPeers = "false";
-			}
-
-			// Write new settings in config.php
-			if (file_exists('config.php')){
-				$conf = file_get_contents('config.php');
-				$conf = preg_replace("/geoPeers = (true|false);/i", 'geoPeers = '.$geoPeers.';', $conf);
-				file_put_contents('config.php', $conf);
-				$message = "Setings succesfully saved";
-			}else{
-				$error = "Config file does not exists";
-			}				
-			$message = "Settings succesfully saved";
-		}
+}elseif($_GET['p'] == 'pastorders') {
+	$days = 1;
+	$maker = '';
+	$taker = '';
+	$snode = '';
+	if(isset($_GET['days'])){
+		$days = $_GET['days'];
 	}
-   $data = array('section' => 'settings', 'title' => 'Settings', 'geoPeers' => Config::PEERS_GEO);
+	if(isset($_GET['maker'])){
+		$maker = $_GET['maker'];
+	}
+	if(isset($_GET['taker'])){
+		$taker = $_GET['taker'];
+	}
+	if(isset($_GET['snode'])){
+		$snode = $_GET['snode'];
+	}
+	$content= createPastOrdersContent($days, $maker, $taker, $snode);
+	$data = array('section' => 'pastorders', 'title' => 'Past Orders', 'content' => $content);
+  
+// DX/XR Wallets Page 
+}elseif($_GET['p'] == 'dxxrwallets') {
+	$content= createDxXrWallets();
+	$data = array('section' => 'dxxrwallets', 'title' => 'DX+XR Wallets', 'content' => $content);
+  
+// XRouter services Page 
+}elseif($_GET['p'] == 'xrservices') {
+	$service = '';
+	$coin = '';
+	$snode = '';
+	if(isset($_GET['service'])){
+		$service = $_GET['service'];
+	}
+	if(isset($_GET['coin'])){
+		$coin = $_GET['coin'];
+	}
+	if(isset($_GET['snode'])){
+		$snode = $_GET['snode'];
+	}
+	$content= createXrServices($snode, $coin, $service);
+	$data = array('section' => 'xrservices', 'title' => 'XRouter Services', 'content' => $content);
+  
+// XCloud services Page 
+}elseif($_GET['p'] == 'xcservices') {
+	$service = '';
+	$snode = '';
+	if(isset($_GET['service'])){
+		$service = $_GET['service'];
+	}
+	if(isset($_GET['snode'])){
+		$snode = $_GET['snode'];
+	}
+	$content= createXcServices($snode, $service);
+	$data = array('section' => 'xcservices', 'title' => 'XCloud Services', 'content' => $content);
 
+	// Trades and fees Page 
+}elseif($_GET['p'] == 'tradesfees') {
+	$days = '';
+	if(isset($_GET['days'])){
+		$days = $_GET['days'];
+	}
+	$content= createTradesAndFees($days);
+	$data = array('section' => 'tradesfees', 'title' => 'Trades and Fees', 'content' => $content);
+
+	// Database update Page 
+}elseif($_GET['p'] == 'dbupdate') {
+	$content= dbupdate(1);
+	$data = array('section' => 'dbupdate', 'title' => 'DB Update', 'content' => $content);
+  
 // About Page	
-}elseif($_GET['p'] == "about") {
-	$data = array('section' => 'about', 'title' => 'About'); 
+}elseif($_GET['p'] == 'about') {
+	$content= dbupdate();
+	$data = array('section' => 'about', 'title' => 'About', 'content' => $content); 
 	
 }else{
 	header('Location: index.php');