+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 Scalaris Monitor Changelog
 ================================
 
+## 0.2.0 - 2021-09-22
+* Extend SQLite3 database to past proposals, servicenodes and services
+
+## 0.1.1 - 2021-08-14 
+* Use SQLite3 for past orders
+* Updated colour scheme
+
 ## 0.1.0 - 2021-05-08 Created from 0.1.1 Blocknet Monitor